feat: make some methods of BrowseFilterService protected
